uses app CRD to discover app namespace
Prior to this change, all resources (e.g ingresses, loadbalancers) would
be created on a single namespace (configured from cli flags).

This change introduces support to the apps.tsuru.io custom resource that
is going to be added to tsuru v1.7. Depending on how tsuru api is
configured, an application service may be created on a default namespace
our on a namespace specific for that app's pool. By fetching this
information from the app CRD we are able to know where to create the
app's resources.