Unify logic for Create and Update LB in a single sync method
Previously we had two different code-paths that could cause a service to
be Updated, and each code path would possibly generate a different
service. Now the logic is unified in a single sync which always Update
the service and creates it if it doesn't exist.

This also fixes a bug where ports would always be removed on Create and
added again on Update possibly with different NodePort values, causing
the LB to be recreated on the cloud provider.